fix: pin to major Node.js and npm versions
Customers are often surprised with new issues when using open version range constraints. They might be trying to fix a bug or release a feature, but because they deploy after the release of Node.js 21, they also end up with a new Node.js version. I recommend pinning to a major version, so that developers get to choose when they want to upgrade. This strategy also gives developers the opportunity to deploy the Node.js version change in isolation.